art: ArtGallery + ArtScrollSync
Port components/art/art-gallery.tsx and art-scroll-sync.tsx → Svelte 5.
ArtGallery watches scroll on #art-scroller and activates a sliding
window of SketchHost instances. ArtScrollSync keeps URL hash in sync
with the visible snap section.
